Refactor processProvidedScenarios for reusability and clarity.
